(only sampling for) multiple passes working
-  checked which textures need to be cleared (now only clearing those)

+++ src/unity/concept/Assets/Testing/MultiPass/Geometry.cginc
@@ -6,8 +6,8 @@ struct ProjectedPoint
     // absolute projected coordinates
     int2 pc;
 
-    // corresponding original coordinates
-    int2 oc;
+    // corresponding original uv coordinates
+    float2 oc;
 
     // projected depth value (non-normalized)
     float d;
